+++ Yeti/AppDelegate/AppDelegate.m
@@ -115,7 +115,7 @@ AppDelegate *MyAppDelegate = nil;
         [UIApplication registerObjectForStateRestoration:(id <UIStateRestoring>)MyFeedsManager restorationIdentifier:NSStringFromClass(FeedsManager.class)];
         [UIApplication registerObjectForStateRestoration:(id <UIStateRestoring>)ArticlesManager.shared restorationIdentifier:NSStringFromClass(ArticlesManager.class)];
         
-        // To test push notifications
+//         To test push notifications
 //        #ifdef DEBUG
 //            dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(1 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
 //                strongify(self);
@@ -123,7 +123,7 @@ AppDelegate *MyAppDelegate = nil;
 //        //        [self openFeed:@(1) article:@(1293968)];  // twitter user
 //        //        [self openFeed:@(1) article:@(1273075)];  // twitter status
 //        //        [self openFeed:@(1) article:@(1149498)];  // reddit
-////                [self openFeed:@(73) article:@(8301134)];
+//                [self openFeed:@(1553) article:@(9542440)];
 ////                [self showArticle:@(1831527)]; // crashing article
 //            });
 //        #endif